+++ NEWS.md
@@ -4,6 +4,7 @@
 
 # Version 1.1.4
 - Flexible temporal resolution of model evaluation
+- Basal area index evaluation
 
 # Version 1.1.3
 - New model evaluation functions